refactor(arch): reorganize codebase using MVC pattern
